Create ESP32 + DHT22 + HiveMQ.ino
This code runs on an ESP32 board. It connects to Wi-Fi and reads temperature and humidity from a DHT22 sensor. Every 5 seconds, it sends this data to the HiveMQ public MQTT server using the topic index/iot/dht22. The data is sent in JSON format. This allows you to monitor the sensor values from anywhere using an MQTT client.